financial_bot/requirements.txt
Flask
flask-cors
flask-login
Flask-SQLAlchemy
sqlalchemy
werkzeug
pandas
yfinance
requests
beautifulsoup4
python-dotenv
langchain_chroma
langchain_groq
langchain_google_genai
langchain_community
markdown
sentence-transformers>=2.2.2

